+++ src/core/state_estimator/ins/ins_sensor_sync.h
@@ -14,8 +14,8 @@ typedef struct {
 
 typedef struct {
 	float timestamp_s;
-	float longitude;
-	float latitude;
+	int32_t longitude;
+	int32_t latitude;
 	float height_msl;
 	float vx_ned;
 	float vy_ned;
@@ -45,12 +45,12 @@ void ins_barometer_sync_buffer_push_from_isr(float height, float height_rate,
 bool ins_barometer_sync_buffer_pop(float *height, float *height_rate);
 
 bool ins_gps_sync_buffer_available(void);
-void ins_gps_sync_buffer_push(float longitude, float latitude, float height_msl,
+void ins_gps_sync_buffer_push(int32_t longitude, int32_t latitude, float height_msl,
                               float vx_ned, float vy_ned, float vz_ned);
-void ins_gps_sync_buffer_push_from_isr(float longitude, float latitude, float height_msl,
+void ins_gps_sync_buffer_push_from_isr(int32_t longitude, int32_t latitude, float height_msl,
                                        float vx_ned, float vy_ned, float vz_ned,
                                        BaseType_t *higher_priority_task_woken);
-bool ins_gps_sync_buffer_pop(float *longitude, float *latitude, float *height_msl,
+bool ins_gps_sync_buffer_pop(int32_t *longitude, int32_t *latitude, float *height_msl,
                              float *vx_ned, float *vy_ned, float *vz_ned);
 
 bool ins_compass_sync_buffer_available(void);
